+++ demos/petstore/src/main/java/module-info.java
@@ -20,7 +20,7 @@
  * Petstore demo.
  */
 module is.codion.framework.demos.petstore {
-	requires is.codion.swing.common.loadtest.ui;
+	requires is.codion.tools.loadtest.ui;
 	requires is.codion.swing.framework.ui;
 
 	exports is.codion.framework.demos.petstore.model